Rename .*FeatureAssertionSpec classes without fusing
Easier one this time: renamed specs which do not have a counterpart w/o
"Feature", thus do not require fusing.
