+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/ListAssertionsSpec.kt
@@ -9,12 +9,12 @@ import ch.tutteli.atrium.specs.withNullableSuffix
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
 
-object ListFeatureAssertionsSpec : Spek({
+object ListAssertionsSpec : Spek({
     include(ListSpec)
     include(IterableSpec)
     include(SequenceSpec)
 }) {
-    object ListSpec : ch.tutteli.atrium.specs.integration.ListFeatureAssertionsSpec(
+    object ListSpec : ch.tutteli.atrium.specs.integration.ListAssertionsSpec(
         feature1<List<Int>, Int, Int>(Expect<List<Int>>::get),
         fun2<List<Int>, Int, Expect<Int>.() -> Unit>(Expect<List<Int>>::get),
         feature1<List<Int?>, Int, Int?>(Expect<List<Int?>>::get).withNullableSuffix(),
