+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/PairAssertionsSpec.kt
@@ -1,12 +1,11 @@
-package ch.tutteli.atrium.api.infix.en_GB
+package ch.tutteli.atrium.api.fluent.en_GB
 
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.specs.fun1
 import ch.tutteli.atrium.specs.notImplemented
 import ch.tutteli.atrium.specs.property
-import ch.tutteli.atrium.specs.testutils.WithAsciiReporter
 
-class PairFeatureAssertionsSpec : ch.tutteli.atrium.specs.integration.PairFeatureAssertionsSpec(
+class PairAssertionsSpec : ch.tutteli.atrium.specs.integration.PairAssertionsSpec(
     property<Pair<String, Int>, String>(Expect<Pair<String, Int>>::first),
     fun1<Pair<String, Int>, Expect<String>.() -> Unit>(Expect<Pair<String, Int>>::first),
     property<Pair<String, Int>, Int>(Expect<Pair<String, Int>>::second),
@@ -16,7 +15,6 @@ class PairFeatureAssertionsSpec : ch.tutteli.atrium.specs.integration.PairFeatur
     property<Pair<String?, Int?>, Int?>(Expect<Pair<String?, Int?>>::second),
     fun1<Pair<String?, Int?>, Expect<Int?>.() -> Unit>(Expect<Pair<String?, Int?>>::second)
 ) {
-    companion object : WithAsciiReporter()
 
     @Suppress("unused", "UNUSED_VALUE")
     private fun ambiguityTest() {
@@ -31,21 +29,21 @@ class PairFeatureAssertionsSpec : ch.tutteli.atrium.specs.integration.PairFeatur
         a3.first
         a4.first
         star.first
-        a1 = a1 first { }
-        a2 = a2 first { }
-        a3 = a3 first { }
-        a4 = a4 first { }
-        star = star first { }
+        a1 = a1.first { }
+        a2 = a2.first { }
+        a3 = a3.first { }
+        a4 = a4.first { }
+        star = star.first { }
 
         a1.second
         a2.second
         a3.second
         a4.second
         star.second
-        a1 = a1 second { }
-        a2 = a2 second { }
-        a3 = a3 second { }
-        a4 = a4 second { }
-        star = star second { }
+        a1 = a1.second { }
+        a2 = a2.second { }
+        a3 = a3.second { }
+        a4 = a4.second { }
+        star = star.second { }
     }
 }
