+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-jvm/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/ZonedDateTimeAssertionsSpec.kt
@@ -1,14 +1,13 @@
-package ch.tutteli.atrium.api.infix.en_GB
+package ch.tutteli.atrium.api.fluent.en_GB
 
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.specs.fun1
 import ch.tutteli.atrium.specs.notImplemented
 import ch.tutteli.atrium.specs.property
-import ch.tutteli.atrium.specs.testutils.WithAsciiReporter
 import java.time.DayOfWeek
 import java.time.ZonedDateTime
 
-class ZonedDateTimeFeatureAssertionsSpec : ch.tutteli.atrium.specs.integration.ZonedDateTimeFeatureAssertionsSpec(
+class ZonedDateTimeAssertionsSpec : ch.tutteli.atrium.specs.integration.ZonedDateTimeAssertionsSpec(
     property<ZonedDateTime, Int>(Expect<ZonedDateTime>::year),
     fun1<ZonedDateTime, Expect<Int>.() -> Unit>(Expect<ZonedDateTime>::year),
     property<ZonedDateTime, Int>(Expect<ZonedDateTime>::month),
@@ -18,22 +17,20 @@ class ZonedDateTimeFeatureAssertionsSpec : ch.tutteli.atrium.specs.integration.Z
     property<ZonedDateTime, DayOfWeek>(Expect<ZonedDateTime>::dayOfWeek),
     fun1<ZonedDateTime, Expect<DayOfWeek>.() -> Unit>(Expect<ZonedDateTime>::dayOfWeek)
 ) {
-    companion object : WithAsciiReporter()
-
     @Suppress("unused", "UNUSED_VALUE")
     private fun ambiguityTest() {
         var a1: Expect<ZonedDateTime> = notImplemented()
 
         a1.year
-        a1 = a1 year { }
+        a1 = a1.year { }
 
         a1.month
-        a1 = a1 month { }
+        a1 = a1.month { }
 
         a1.dayOfWeek
-        a1 = a1 dayOfWeek { }
+        a1 = a1.dayOfWeek { }
 
         a1.day
-        a1 = a1 day { }
+        a1 = a1.day { }
     }
 }
