+++ misc/deprecated/apis/fluent-en_GB-jdk8/atrium-api-fluent-en_GB-jdk8-jvm/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/jdk8/ZonedDateTimeAssertionsSpec.kt
@@ -10,7 +10,7 @@ import ch.tutteli.atrium.specs.property
 import java.time.DayOfWeek
 import java.time.ZonedDateTime
 
-class ZonedDateTimeFeatureAssertionsSpec : ch.tutteli.atrium.specs.integration.ZonedDateTimeFeatureAssertionsSpec(
+class ZonedDateTimeAssertionsSpec : ch.tutteli.atrium.specs.integration.ZonedDateTimeAssertionsSpec(
     property<ZonedDateTime, Int>(Expect<ZonedDateTime>::year),
     fun1<ZonedDateTime, Expect<Int>.() -> Unit>(Expect<ZonedDateTime>::year),
     property<ZonedDateTime, Int>(Expect<ZonedDateTime>::month),
